--- src/main/java/com/example/atm/base/exception/ExceptionCode.java
@@ -1,11 +1,0 @@
-package com.example.atm.base.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ExceptionCode {
-    public HttpStatus getHttpStatus();
-
-    public String getCause();
-
-    public String getName();
-}
