--- src/main/java/com/example/atm/base/exception/GlobalException.java
@@ -1,14 +1,0 @@
-package com.example.atm.base.exception;
-
-import lombok.Getter;
-
-@Getter
-public class GlobalException extends RuntimeException {
-
-    ExceptionCode exceptionCode;
-
-    public GlobalException(ExceptionCode exceptionCode) {
-        super(exceptionCode.getCause());
-        this.exceptionCode = exceptionCode;
-    }
-}
