--- src/main/java/com/example/atm/base/jwt/JwtProperties.java
@@ -1,14 +1,0 @@
-package com.example.atm.base.jwt;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-@Data
-@Configuration
-@ConfigurationProperties(prefix = "jwt")
-public class JwtProperties {
-    private String issuer;
-    private String secretKey;
-    private Long expiration;
-}
