--- src/main/java/com/example/atm/bounded_context/user/dto/UserInfoDto.java
@@ -1,32 +1,0 @@
-package com.example.atm.bounded_context.user.dto;
-
-import com.example.atm.bounded_context.user.entity.Gender;
-import com.example.atm.bounded_context.user.entity.User;
-
-import java.util.Optional;
-
-public record UserInfoDto(
-        Long id,
-        String email,
-        String nickname,
-        String profileImgUrl,
-        Gender gender,
-        Integer point,
-        Long socialId,
-        Long mateId
-) {
-    public static UserInfoDto fromEntity(User user) {
-        return new UserInfoDto(
-                user.getId(),
-                user.getEmail(),
-                user.getNickname(),
-                user.getProfileImgUrl(),
-                user.getGender(),
-                user.getPoint(),
-                user.getSocialId(),
-                Optional.ofNullable(user.getMate())
-                        .map(User::getId)
-                        .orElse(null)
-        );
-    }
-}
