--- src/main/java/com/example/atm/bounded_context/user/entity/Gender.java
@@ -1,5 +1,0 @@
-package com.example.atm.bounded_context.user.entity;
-
-public enum Gender {
-    MALE, FEMALE
-}
