+++ src/main/java/com/example/holing/base/exception/ExceptionCode.java
@@ -1,0 +1,11 @@
+package com.example.holing.base.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface ExceptionCode {
+    public HttpStatus getHttpStatus();
+
+    public String getCause();
+
+    public String getName();
+}
