+++ src/main/java/com/example/holing/base/exception/ExceptionResponse.java
@@ -1,0 +1,17 @@
+package com.example.holing.base.exception;
+
+import java.time.LocalDateTime;
+
+public record ExceptionResponse(
+        LocalDateTime timeStamp,
+        String name,
+        String cause
+) {
+    public static ExceptionResponse fromEntity(ExceptionCode exceptionCode) {
+        return new ExceptionResponse(
+                LocalDateTime.now(),
+                exceptionCode.getName(),
+                exceptionCode.getCause()
+        );
+    }
+}
