+++ src/main/java/com/example/holing/base/exception/GlobalException.java
@@ -1,0 +1,14 @@
+package com.example.holing.base.exception;
+
+import lombok.Getter;
+
+@Getter
+public class GlobalException extends RuntimeException {
+
+    ExceptionCode exceptionCode;
+
+    public GlobalException(ExceptionCode exceptionCode) {
+        super(exceptionCode.getCause());
+        this.exceptionCode = exceptionCode;
+    }
+}
