+++ src/main/java/com/example/holing/bounded_context/auth/dto/OAuthTokenInfoDto.java
@@ -1,0 +1,11 @@
+package com.example.holing.bounded_context.auth.dto;
+
+public record OAuthTokenInfoDto(
+        String accessToken,
+        String refreshToken,
+        String scope
+) {
+    public static OAuthTokenInfoDto of(String accessToken, String refreshToken, String scope) {
+        return new OAuthTokenInfoDto(accessToken, refreshToken, scope);
+    }
+}
