+++ src/main/java/com/example/holing/bounded_context/schedule/dto/ScheduleRequestDto.java
@@ -1,0 +1,21 @@
+package com.example.holing.bounded_context.schedule.dto;
+
+import com.example.holing.bounded_context.schedule.entity.Schedule;
+
+import java.time.LocalDateTime;
+
+public record ScheduleRequestDto(
+        String title,
+        String place,
+        LocalDateTime startAt,
+        LocalDateTime finishAt
+) {
+    public Schedule toEntity() {
+        return Schedule.builder()
+                .title(title)
+                .place(place)
+                .startAt(startAt)
+                .finishAt(finishAt)
+                .build();
+    }
+}
