+++ src/main/java/com/example/holing/bounded_context/user/entity/Gender.java
@@ -1,0 +1,5 @@
+package com.example.holing.bounded_context.user.entity;
+
+public enum Gender {
+    MALE, FEMALE
+}
