--- src/test/java/com/example/atm/AtmApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.atm;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AtmApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
