+++ src/test/java/com/example/holing/HolingApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.holing;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HolingApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
